[SW-24541] Use different way to include xmldom dependency
**Summary**
It turned out that fixing the syntax of the git tag wasn't enough. Fortunately, according to https://github.com/auth0/xmldom, it appears there is a mapping between the old way and a normal way of including the dependency.

**Test Plan**
This results in the same commit hash being downloaded for the dependency as before. Unfortunately, this repository does not include a package-lock.json which can be used to verify that in the pull request.